fix(wilro): lora_alpha declared int, written as float — checkpoints would not load
train_wilro defaults lora_alpha to 2 x rank and casts to float, so
--lora_rank 64 writes 128.0 into config.json. WilroConfig declared the
field as int, and draccus refuses that on the way back in:

    DecodingError: `lora_alpha`: Couldn't parse '128.0' into an int

A run therefore trained to completion and then could not be evaluated --
the failure surfaces at PreTrainedConfig.from_pretrained, nowhere near the
trainer that wrote it.

Widened to float, which is also the honest type: LoRALinear uses alpha only
as the ratio alpha/rank. Verified through the real load path
(PreTrainedConfig.from_pretrained) that 32, 64, 128 and 128.0 all decode,
so checkpoints written before this -- whose value is a plain int -- still
load, and the one that triggered this now loads without being edited.

+++ src/models/wilro/wilro_config.py
@@ -213,7 +213,13 @@ class WilroConfig(PreTrainedConfig):
     # robot-specific instructions and spatial grounding.
     # Base weights stay frozen; only LoRA params are trainable.
     lora_rank: int = 16
-    lora_alpha: int = 32
+    # float, not int: LoRALinear uses it only as the ratio alpha/rank, and
+    # the trainer's default of 2 x rank is written out as e.g. 128.0. Declared
+    # as int, draccus refused to load any checkpoint carrying such a value --
+    #   DecodingError: `lora_alpha`: Couldn't parse '128.0' into an int
+    # so a run trained fine and then could not be evaluated. Widening also
+    # reads the older checkpoints, whose value is a plain 32.
+    lora_alpha: float = 32.0
     lora_dropout: float = 0.05
     lora_target_modules: list = field(default_factory=lambda: ["q_proj", "v_proj"])
     vision_lora_num_layers: int = 8  # Last 8 layers of SigLIP ViT get LoRA
